Update README.md
In this project, my team wanted to investigate the scarceness of NBA positions, and whether or not team general managers should invest in these scarce positions based on salary data. My role in this team was to gather data from an espn website, clean and analyze for summary statistics, particularly in salary data.